GUI Update 3
Added additional widgets for team testing as well as set up the basis of
storing and saving settings during the gameplay within a .pixel file
format, this is read using the pickle module and PIL is being used to
allow for images to be read. GIF image control also removed as it is
currently unrequired. Also Tom Starling has created 3 different playing
terrains for the user to use. - William Read