[Help] Bugged Canvas on bubble_sort
I can't get this bubble shit working. It's sorting the list and adding a new item to the list whenever you press the button like when you will be collecting turret parts each part will have an ID and it will be added to the main list. So the sort is working but I can't get the Canvas actually show it...