Final Video Edits
This system is able to capture in-game screen shots and add the
available features to them as well as alerting the user once they have
been taken. It also has the ability to film .AVI based using a
compilation of image capture. Audio still has issues capturing, and hope
to include Youtube AI to allow for immediate uploads.